Started weather exporter app
This adds a docker-compose.yml to define our app
Adds a mongo instance and creates a mongo database
Outlines the next work to be completed
