+++ Dockerfile
@@ -1,7 +1,7 @@
 FROM python:3
 ARG LISTEN_PORT="3333"
-ADD helloworld.py /
-RUN pip install flask
+ADD frontend.py /
+RUN pip install flask pymongo
 EXPOSE ${LISTEN_PORT}
 ENTRYPOINT [ "python" ]
-CMD [ "helloworld.py" ]
+CMD [ "frontend.py" ]
