+++ docker-compose.yml
@@ -1,0 +1,19 @@
+services:
+
+  frontend:
+    build: .
+    ports:
+    - 8888:3333
+    environment:
+      HTTP_PORT: 3333
+      DB_NAME: weather_stats
+    depends_on: [ "mongodb" ]
+
+  mongodb:
+    image: mongo:latest
+    volumes:
+    - mongo:/data/db
+    expose: [ 27017 ]
+
+volumes:
+  mongo:
